Move test to right component
